feat: add Definition CRUD

+++ api/views/definition.py
@@ -14,23 +14,7 @@ from api.services.user import UserService
 class DefinitionView(APIView):
     @staticmethod
     def post(request):
-        if not request.user.is_authenticated:
-            return Response(status=status.HTTP_401_UNAUTHORIZED)
-
-        if not request.user.is_staff:
-            return Response(status=status.HTTP_403_FORBIDDEN)
-
-        log.debug(f'{request.user.is_staff=}')
-
-        serializer = DefinitionSerializer(data=request.data)
-
-        if not serializer.is_valid():
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-        definition = DefinitionService.create(serializer)
-        serializer = DefinitionSerializer(definition)
-
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
 
 class SingleDefinitionView(APIView):
@@ -38,6 +22,9 @@ class SingleDefinitionView(APIView):
     def get(request, pk: int):
         if not DefinitionService.exists(pk):
             return Response(status=status.HTTP_404_NOT_FOUND)
+
+        log.debug(f"{UserService.can_see_non_validated_entries(request.user)=}")
+        log.debug(f"{DefinitionService.is_parent_validated(pk)=}")
 
         if (
                 not UserService.can_see_non_validated_entries(request.user) and
@@ -74,6 +61,12 @@ class SingleDefinitionView(APIView):
     def delete(request, pk: int):
         if not DefinitionService.exists(pk):
             return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if not request.user.is_authenticated:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
+        if not request.user.is_staff:
+            return Response(status=status.HTTP_403_FORBIDDEN)
 
         DefinitionService.delete(pk)
         return Response(status=status.HTTP_204_NO_CONTENT)
